SHA256 hashing implemented and version updated
In the file `SendlixHandler.cs` a new dependency to `System.Security.Cryptography` has been added to enable SHA256 hashing. Changed the type of the `authCache` from `ConcurrentDictionary<long, AuthResponse>` to `ConcurrentDictionary<string, AuthResponse>` to use the hashed key. The code for calculating the hashed secret has been added and the logic for checking and adding entries to the cache has been updated.

Additionally, the version of the project in the file `Sendlix.Smpt.Relay.csproj` has been updated from `1.0.0` to `1.1.1`.